# frontend/public/openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Tori API",
    "version": "2026-07-01",
    "description": "Managed recurring billing engine built natively on Nomba. Start a subscription with one API call. Tori handles everything after — dunning, renewals, recovery, ledger, and customer self-service.\n\nThis spec documents the integration surface only: the Platform API you call from your own server, the webhooks Tori sends back, and the Customer Portal your customers can authenticate into directly. It does not document the Tori dashboard's own internal endpoints — those exist to power the dashboard UI, not for third-party integration.",
    "contact": {
      "name": "Tori Support",
      "email": "support@tori.ng",
      "url": "https://frontend-production-e3be.up.railway.app/docs"
    }
  },
  "servers": [
    {
      "url": "https://api-production-3847.up.railway.app",
      "description": "Production (use tori_live_ keys)"
    },
    {
      "url": "https://api-production-3847.up.railway.app",
      "description": "Sandbox (use tori_test_ keys)"
    }
  ],
  "security": [],
...
  "paths": {
    "/v1/oauth/token": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Get an access token",
        "description": "Exchanges OAuth client credentials for a short-lived bearer token, usable anywhere an X-API-Key is accepted. An alternative to a static API key — use whichever fits your integration; both authenticate the same Platform API surface.",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "grant_type",
                  "client_id",
                  "client_secret"
                ],
                "properties": {
                  "grant_type": {
                    "type": "string",
                    "enum": [
                      "client_credentials"
                    ]
                  },
                  "client_id": {
                    "type": "string",
                    "example": "oauth_client_9913c788057e1c1f0310cd79"
                  },
                  "client_secret": {
                    "type": "string",
                    "example": "oauth_secret_a8f1454315f7bf5f204a90ad7ffe694c"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Token issued",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "access_token": {
                          "type": "string"
                        },
                        "token_type": {
                          "type": "string",
                          "enum": [
                            "Bearer"
                          ]
                        },
                        "expires_in": {
                          "type": "integer",
                          "example": 1800
                        },
                        "mode": {
                          "type": "string",
                          "enum": [
                            "live",
                            "test"
                          ]
                        }
                      }
                    },
                    "meta": {
                      "$ref": "#/components/schemas/Meta"
                    }
                  }
                },
                "example": {
                  "data": {
                    "access_token": "tori_oauth_a8f1454315f7bf5f204a90ad7ffe694c34ff1c382b66edf061a6b96027ab92cd",
                    "token_type": "Bearer",
                    "expires_in": 1800,
                    "mode": "live"
                  },
                  "meta": {
                    "request_id": "5f1b2c3d-0000-4000-8000-000000000000",
                    "api_version": "2026-07-01"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": {
                    "code": "invalid_body",
                    "message": "request body is not valid JSON"
                  },
                  "meta": {
                    "request_id": "5f1b2c3d-0000-4000-8000-000000000000",
                    "api_version": "2026-07-01"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": {
                    "code": "unauthorised",
                    "message": "invalid client credentials"
                  },
                  "meta": {
                    "request_id": "5f1b2c3d-0000-4000-8000-000000000000",
                    "api_version": "2026-07-01"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": {
                    "code": "not_found",
                    "message": "the requested resource does not exist"
                  },
                  "meta": {
                    "request_id": "5f1b2c3d-0000-4000-8000-000000000000",
                    "api_version": "2026-07-01"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/v1/auth/login": {
      "post": {
        "tags": [
          "Authentication"
        ],
        "summary": "Log in as a tenant operator",
        "description": "Dashboard/operator login — not used by customer-facing integrations. Returns a JWT pair scoped to the tenant.",
        "security": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "email",
                  "password"
                ],
                "properties": {
                  "email": {
                    "type": "string",
                    "format": "email"
                  },
                  "password": {
                    "type": "string",
                    "format": "password"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Logged in",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "object",
                      "properties": {
                        "access_token": {
                          "type": "string"
                        },
                        "refresh_token": {
                          "type": "string"
                        },
                        "token_type": {
                          "type": "string",
                          "enum": [
                            "Bearer"
                          ]
                        },
                        "email_verified": {
                          "type": "boolean"
                        }
                      }
                    },
                    "meta": {
                      "$ref": "#/components/schemas/Meta"
                    }
                  }
                },
                "example": {
                  "data": {
                    "access_token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...",
                    "refresh_token": "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...",
                    "token_type": "Bearer",
                    "email_verified": true
                  },
                  "meta": {
                    "request_id": "5f1b2c3d-0000-4000-8000-000000000000",
                    "api_version": "2026-07-01"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": {
                    "code": "invalid_body",
                    "message": "request body is not valid JSON"
                  },
                  "meta": {
                    "request_id": "5f1b2c3d-0000-4000-8000-000000000000",
                    "api_version": "2026-07-01"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": {
                    "code": "unauthorised",
                    "message": "invalid credentials"
                  },
                  "meta": {
                    "request_id": "5f1b2c3d-0000-4000-8000-000000000000",
                    "api_version": "2026-07-01"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": {
                    "code": "not_found",
                    "message": "the requested resource does not exist"
                  },
                  "meta": {
                    "request_id": "5f1b2c3d-0000-4000-8000-000000000000",
                    "api_version": "2026-07-01"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/platform/subscriptions": {
      "get": {
        "tags": [
          "Platform API — Subscriptions"
        ],
        "summary": "List subscriptions",
        "description": "Returns subscriptions for your account, newest first, offset-paginated.",
        "security": [
          {
            "apiKeyAuth": []
          },
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "X-API-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string"
            },
            "example": "tori_live_..."
          },
          {
            "name": "limit",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 20
            },
            "description": "Max results per page"
          },
          {
            "name": "offset",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 0
            },
            "description": "Results to skip"
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string"
            },
            "description": "Filter by subscription status"
          }
        ],
        "responses": {
          "200": {
            "description": "Subscriptions returned",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/Subscription"
                      }
                    },
                    "pagination": {
                      "type": "object",
                      "properties": {
                        "has_more": {
                          "type": "boolean"
                        },
                        "total": {
                          "type": "integer"
                        }
                      }
                    },
                    "meta": {
                      "$ref": "#/components/schemas/Meta"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": {
                    "code": "invalid_body",
                    "message": "request body is not valid JSON"
                  },
                  "meta": {
                    "request_id": "5f1b2c3d-0000-4000-8000-000000000000",
                    "api_version": "2026-07-01"
                  }
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": {
                    "code": "unauthorised",
                    "message": "invalid or missing API key"
                  },
                  "meta": {
                    "request_id": "5f1b2c3d-0000-4000-8000-000000000000",
                    "api_version": "2026-07-01"
                  }
                }
              }
            }
          },
          "404": {
            "description": "Not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                },
                "example": {
                  "error": {
                    "code": "not_found",
                    "message": "the requested resource does not exist"
                  },
                  "meta": {
                    "request_id": "5f1b2c3d-0000-4000-8000-000000000000",
                    "api_version": "2026-07-01"
                  }
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "apiKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Your Tori API key. Use tori_live_ for production, tori_test_ for sandbox."
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "OAuth 2.0 Bearer token from POST /v1/oauth/token, or dashboard JWT from POST /v1/auth/login"
      },
      "portalAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "Portal JWT from POST /v1/portal/auth/verify-otp"
      }
    },
    "schemas": {
      "Meta": {
        "type": "object",
        "properties": {
          "request_id": {
            "type": "string",
            "format": "uuid"
          },
          "api_version": {
            "type": "string"
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              }
            }
          },
          "meta": {
            "$ref": "#/components/schemas/Meta"
          }
        }
      },
...
    }
  },
...
}